requirements.txt
asgiref==3.7.2
Django==4.2.6
fpdf==1.7.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
sqlparse==0.4.4
typing_extensions==4.8.0
